active_interview_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
coverage==7.6.12
cssbeautifier==1.15.4
distro==1.9.0
Django==4.2.19
django-bootstrap-v5==1.0.11
django-filter==25.1
djangorestframework==3.15.2
EditorConfig==0.17.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsbeautifier==1.15.4
json5==0.10.0
Markdown==3.7
mccabe==0.7.0
openai==1.65.4
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
PySocks==1.7.1
PyYAML==6.0.2
regex==2024.11.6
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
